list_cart+cartAdapter+moved
